chores: localUser 딜리트 기능 수정 및 소켓 수정

+++ src/routes/controllers/user.Controller.js
@@ -91,7 +91,7 @@ exports.deleteUser = async (req, res, next) => {
   const { email } = req.body;
 
   try {
-    await User.deleteOne({ email });
+    await User.deleteOne({ uid: email });
 
     res.status(204).send({ result: "delete" });
   } catch (err) {
